pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["DeveloperInal <166642110+DeveloperInal@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.13.0"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"
loguru = "^0.7.2"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
